Move plot function to utils
